ajuste na estrutura de pastas da tela home (principalmente) + renames para melhorar a organização do projeto
