+++ ui/src/components/landing/landing-screen-content/LandingScreenContentCP.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import TextCP from '../../../common/components/text/TextCP'
 import HintoTextCP from '../../../common/components/hinto-text/HintoTextCP'
-import LandingCardCP from '../landing-card/LandingCardCP'
+import LandingMovieCardCP from '../landing-movie-card/LandingMovieCardCP'
 import LandingTopTextCP from '../landing-top-text/LandingTopTextCP'
 import { LandingScreenContentMock } from '../../../mocks/LandingScreenContentMock'
 import { ILandingScreenResponseDTO } from '../../../interfaces/dtos/response/ILandingScreenResponseDTO'
@@ -25,7 +25,7 @@ export default function LandingScreenContentCP(): JSX.Element {
       </LandingTopTextSCP>
 
       <MainCardWrapperSCP>
-        <LandingCardCP
+        <LandingMovieCardCP
           main={true}
           synopsis={MOCK[0].synopsis}
           movieTitle={MOCK[0].movieTitle}
@@ -49,19 +49,19 @@ export default function LandingScreenContentCP(): JSX.Element {
         </BottomTextWrapperSCP>
 
         <DefaultCardsWrapperSCP>
-          <LandingCardCP
+          <LandingMovieCardCP
             synopsis={MOCK[1].synopsis}
             movieTitle={MOCK[1].movieTitle}
             urlImage={MOCK[1].urlImage}
           />
 
-          <LandingCardCP
+          <LandingMovieCardCP
             synopsis={MOCK[2].synopsis}
             movieTitle={MOCK[2].movieTitle}
             urlImage={MOCK[2].urlImage}
           />
 
-          <LandingCardCP
+          <LandingMovieCardCP
             synopsis={MOCK[3].synopsis}
             movieTitle={MOCK[3].movieTitle}
             urlImage={MOCK[3].urlImage}
